feature flag timer tests that need alloc
